+++ R/percentGenes.R
@@ -2,8 +2,8 @@
 #'
 #' This function the proportion V or J genes used by 
 #' grouping variables. This function only quantifies
-#' single gene loci for indicated \strong{chain}. For 
-#' examining VJ pairing, please see \code{\link{percentVJ}}.
+#' single gene loci for indicated **chain**. For 
+#' examining VJ pairing, please see [percentVJ()].
 #'
 #' @examples
 #' #Making combined contig data
@@ -14,8 +14,8 @@
 #'              chain = "TRB", 
 #'              gene = "Vgene")
 #' 
-#' @param input.data The product of \code{\link{combineTCR}}, 
-#' \code{\link{combineBCR}}, or \code{\link{combineExpression}}.
+#' @param input.data The product of [combineTCR()], 
+#' [combineBCR()], or [combineExpression()].
 #' @param chain "TRA", "TRB", "TRG", "TRG", "IGH", "IGL".
 #' @param gene "V", "D" or "J"
 #' @param group.by The variable to use for grouping
@@ -23,7 +23,7 @@
 #' to plot groups in order
 #' @param exportTable Returns the data frame used for forming the graph.
 #' @param palette Colors to use in visualization - input any 
-#' \link[grDevices]{hcl.pals}.
+#' [hcl.pals][grDevices::hcl.pals].
 #' @import ggplot2
 #' @importFrom stringr str_split str_sort 
 #' @importFrom reshape2 melt