+++ R/utils.R
@@ -13,10 +13,10 @@
   return(data.frame)
 }
 
-#Use to shuffle between chains Qile: the NA handling here *might* be related to the unnamed combineTCR bugs from the new rcpp con.df construction
+#Use to shuffle between chains
 #' @importFrom stringr str_split
 #' @keywords internal
-#' @author Ye-Lin Son Nick Borcherding
+#' @author Ye-Lin Son, Nick Borcherding
 .off.the.chain <- function(dat, chain, cloneCall, check = TRUE) {
   chain1 <- toupper(chain) #to just make it easier
   if (chain1 %in% c("TRA", "TRG", "IGH")) {
@@ -340,7 +340,7 @@
     return(x)
 }
 
-# helper for .theCall
+# helper for .theCall # Qile: on second thought - converting to x to lowercase may be a bad idea...
 .convertClonecall <- function(x) {
 
   clonecall_dictionary <- hash::hash(
@@ -587,50 +587,15 @@ is_df_or_list_of_df <- function(x) {
   dictionary[dictionary == "None"] <- NA
   dictionary$v.gene <- stringr::str_split(dictionary[,gene], "[.]", simplify = TRUE)[,1]
   tmp <- na.omit(unique(dictionary[,c(chain, "v.gene")]))
+
   #chunking the sequences for distance by v.gene
-  
-  edge.list <- lapply(str_sort(na.omit(unique(dictionary$v.gene)), numeric = T), function(v_gene) {
-    filtered_df <- dplyr::filter(dictionary, v.gene == v_gene)
-    nucleotides <- filtered_df[[chain]]
-    nucleotides <- sort(unique(str_split(nucleotides, ";", simplify = TRUE)[,1]))
-    
-    if (length(nucleotides) <= 1) return(NULL)
-    
-    results <- list()
-    # Only iterate until the second last element to avoid the issue
-    for (i in 1:(length(nucleotides) - 1)) {
-      for (j in (i + 1):length(nucleotides)) {
-        # Check based on length difference feasibility
-        if (abs(nchar(nucleotides[i]) - nchar(nucleotides[j])) > max(nchar(nucleotides[i]), nchar(nucleotides[j])) * (1 - threshold)) {
-          next
-        }
-        
-        distance <- stringdist::stringdist(nucleotides[i], nucleotides[j], method = "lv")
-        normalized_distance <- 1 - distance / mean(c(nchar(nucleotides[i]), nchar(nucleotides[j])))
-        
-        if (normalized_distance >= threshold) {
-          results[[length(results) + 1]] <- data.frame(
-            from = nucleotides[i],
-            to = nucleotides[j],
-            distance = normalized_distance
-          )
-        }
-      }
-    }
-    
-    do.call(rbind, results)
-  })
-  
-  edge.list <- do.call(rbind, edge.list)
-  
+  edge.list <- .createBcrEdgeListDf(dictionary, chain, threshold)
+
   if(exportGraph) {
-    if(!is.null(edge.list)) {
-      graph <- graph_from_edgelist(as.matrix(edge.list)[,c(1,2)])
-      
-    } else {
-      graph <- NULL
+    if (is.null(edge.list)) {
+      return(NULL)
     }
-    return(graph)
+    return(graph_from_edgelist(as.matrix(edge.list)[, c(1, 2)]))
   }
   
   if (!is.null(dim(edge.list))) { 
@@ -657,3 +622,39 @@ is_df_or_list_of_df <- function(x) {
   return(output)
 }
 
+#' create an edge list dataframe from clustering BCR data with edit distance
+#'
+#' This is a helper for the internal [.lvCompare()] that constructs a directed
+#' graph edge list as a dataframe, with weights being the normalized edit
+#' distance between two v genes.
+#'
+#' @param dictionary row binded loadContigs output with a column `v.gene`
+#' @param chain string. Which v gene type.
+#' @param threshold single numeric between 0 and 1
+#'
+#' @return A data.frame with the columns `to`, `from`, `distance`. To and from
+#' represents a directed edge between two v genes, and the distance being the
+#' normalized edit distance. Those that exceed the threshold are filtered.
+#
+#' @keywords internal
+#' @noRd
+.createBcrEdgeListDf <- function(dictionary, chain, threshold) {
+
+  vGenes <- str_sort(na.omit(unique(dictionary$v.gene)), numeric = TRUE)
+
+  edge.list <- lapply(vGenes, function(v_gene) {
+    filtered_df <- dplyr::filter(dictionary, v.gene == v_gene)
+    nucleotides <- filtered_df[[chain]]
+    nucleotides <- sort(unique(str_split(nucleotides, ";", simplify = TRUE)[, 1]))
+    if (length(nucleotides) <= 1) return(NULL)
+    out <- rcppGetSigSequenceEditDistEdgeListDf(nucleotides, threshold)
+    #print(out)
+    out
+  })
+
+  edgeListDf <- do.call(rbind, edge.list)
+  if (!is.null(edgeListDf) && nrow(edgeListDf) == 0) {
+    return(NULL)
+  }
+  edgeListDf
+}